fix: make release publishing tag-driven
