+++ .github/workflows/publish.yml
@@ -6,40 +6,6 @@ on:
 
 jobs:
   publish:
-    if: ${{ !github.event.release.prerelease }}
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      id-token: write # Required for npm trusted publishing
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v5
-
-      - name: Setup Node
-        uses: actions/setup-node@v4
-        with:
-          node-version: 24
-          registry-url: 'https://registry.npmjs.org'
-
-      - name: Setup Bun
-        uses: oven-sh/setup-bun@v2
-        with:
-          bun-version: 1.3.11
-
-      - name: Check dependency release age
-        run: bun run security:release-age
-
-      - name: Install dependencies
-        run: bun install --frozen-lockfile
-
-      - name: Build packages
-        run: bun run build
-
-      - name: Publish to npm
-        run: bunx changeset publish
-
-  publish-rc:
-    if: ${{ github.event.release.prerelease }}
     runs-on: ubuntu-latest
     permissions:
       contents: read
@@ -67,11 +33,20 @@ jobs:
       - name: Install dependencies
         run: bun install --frozen-lockfile
 
-      - name: Check RC versions
-        run: bun scripts/check-prerelease-release.ts
+      - name: Check release versions
+        env:
+          PRERELEASE_TAG: rc
+          RELEASE_PRERELEASE: ${{ github.event.release.prerelease }}
+          RELEASE_TAG: ${{ github.event.release.tag_name }}
+        run: bun scripts/check-release.ts
 
       - name: Build packages
         run: bun run build
 
+      - name: Publish to npm
+        if: ${{ !github.event.release.prerelease }}
+        run: bunx changeset publish
+
       - name: Publish RC to npm
+        if: ${{ github.event.release.prerelease }}
         run: bunx changeset publish --tag rc
